requirements.txt
asgiref==3.2.10
Django==3.1.2
django-quill-editor==0.1.16
install==1.3.4
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.4.1
Unipath==1.1